requirements.txt
slack-bolt>=1.18.0
python-telegram-bot>=20.0
pyyaml>=6.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
